+++ src/App.js
@@ -1,30 +1,14 @@
-import './App.css';
-import MainPage from './MainPage';
-import LoginPage from './LoginPage'
-import { Link } from 'react-router-dom'
-import classNames from 'classnames'
+import MainPage from './pages/Main'
+import LoginPage from './pages/Login'
 import { Routes, Route } from 'react-router-dom'
-function App() {
+
+function App () {
   return (
-    <div className="App">
-      Hello World!
-      
-      <Routes>
-          <Route
-            exact
-            path='/'
-            element={<MainPage />}
-          />
-          <Route
-            path='login'
-            element={<LoginPage />}
-          />
-        </Routes>
-      {/* <MainPage></MainPage> */}
-      
-    </div>
-    
-  );
+    <Routes>
+      <Route exact path='/' element={<MainPage />} />
+      <Route path='login' element={<LoginPage />} />
+    </Routes>
+  )
 }
 
 export default App;
