+++ src/components/CourseThumbnail.js
@@ -1,0 +1,3 @@
+export default function CourseThumbnail(props) {
+    
+};
